Using Retrofit to make REST API calls and remove unnecessary boilerplate code for API calls

+++ app/src/main/java/app/udacity/android/cn/popularmovies/MovieDetailFragment.java
@@ -49,7 +49,7 @@ public class MovieDetailFragment extends Fragment {
 
         String originalTitle = mMovie.getOriginalTitle();
         String overview = mMovie.getOverview();
-        int voteAvg = mMovie.getVoteAvg();
+        float voteAvg = mMovie.getVoteAvg();
         String releaseDtStr = mMovie.getReleaseDt();
 
         //Deprecated, but needed to support older versions
